+++ pkg/block/fetcher_test.go
@@ -22,7 +22,6 @@ import (
 	"github.com/oklog/ulid"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
 	promtest "github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/prometheus/prometheus/pkg/relabel"
 	"github.com/prometheus/prometheus/tsdb"
@@ -34,6 +33,25 @@ import (
 	"github.com/thanos-io/thanos/pkg/testutil"
 	"gopkg.in/yaml.v2"
 )
+
+func newTestFetcherMetrics() *fetcherMetrics {
+	return &fetcherMetrics{
+		synced:   extprom.NewTxGaugeVec(nil, prometheus.GaugeOpts{}, []string{"state"}),
+		modified: extprom.NewTxGaugeVec(nil, prometheus.GaugeOpts{}, []string{"modified"}),
+	}
+}
+
+type ulidFilter struct {
+	ulidToDelete *ulid.ULID
+}
+
+func (f *ulidFilter) Filter(_ context.Context, metas map[ulid.ULID]*metadata.Meta, synced *extprom.TxGaugeVec, incompleteView bool) error {
+	if _, ok := metas[*f.ulidToDelete]; ok {
+		synced.WithLabelValues("filtered").Inc()
+		delete(metas, *f.ulidToDelete)
+	}
+	return nil
+}
 
 func ULID(i int) ulid.ULID { return ulid.MustNew(uint64(i), nil) }
 
@@ -57,12 +75,8 @@ func TestMetaFetcher_Fetch(t *testing.T) {
 
 		var ulidToDelete ulid.ULID
 		r := prometheus.NewRegistry()
-		f, err := NewMetaFetcher(log.NewNopLogger(), 20, bkt, dir, r, func(_ context.Context, metas map[ulid.ULID]*metadata.Meta, synced GaugeLabeled, _ bool) error {
-			if _, ok := metas[ulidToDelete]; ok {
-				synced.WithLabelValues("filtered").Inc()
-				delete(metas, ulidToDelete)
-			}
-			return nil
+		f, err := NewMetaFetcher(log.NewNopLogger(), 20, bkt, dir, r, []MetadataFilter{
+			&ulidFilter{ulidToDelete: &ulidToDelete},
 		})
 		testutil.Ok(t, err)
 
@@ -337,9 +351,10 @@ func TestLabelShardedMetaFilter_Filter_Basic(t *testing.T) {
 		ULID(6): input[ULID(6)],
 	}
 
-	synced := promauto.With(nil).NewGaugeVec(prometheus.GaugeOpts{}, []string{"state"})
-	testutil.Ok(t, f.Filter(ctx, input, synced, false))
-	testutil.Equals(t, 3.0, promtest.ToFloat64(synced.WithLabelValues(labelExcludedMeta)))
+	m := newTestFetcherMetrics()
+	testutil.Ok(t, f.Filter(ctx, input, m.synced, false))
+
+	testutil.Equals(t, 3.0, promtest.ToFloat64(m.synced.WithLabelValues(labelExcludedMeta)))
 	testutil.Equals(t, expected, input)
 
 }
@@ -434,10 +449,11 @@ func TestLabelShardedMetaFilter_Filter_Hashmod(t *testing.T) {
 			}
 			deleted := len(input) - len(expected)
 
-			synced := promauto.With(nil).NewGaugeVec(prometheus.GaugeOpts{}, []string{"state"})
-			testutil.Ok(t, f.Filter(ctx, input, synced, false))
+			m := newTestFetcherMetrics()
+			testutil.Ok(t, f.Filter(ctx, input, m.synced, false))
+
 			testutil.Equals(t, expected, input)
-			testutil.Equals(t, float64(deleted), promtest.ToFloat64(synced.WithLabelValues(labelExcludedMeta)))
+			testutil.Equals(t, float64(deleted), promtest.ToFloat64(m.synced.WithLabelValues(labelExcludedMeta)))
 
 		})
 
@@ -497,9 +513,10 @@ func TestTimePartitionMetaFilter_Filter(t *testing.T) {
 		ULID(4): input[ULID(4)],
 	}
 
-	synced := promauto.With(nil).NewGaugeVec(prometheus.GaugeOpts{}, []string{"state"})
-	testutil.Ok(t, f.Filter(ctx, input, synced, false))
-	testutil.Equals(t, 2.0, promtest.ToFloat64(synced.WithLabelValues(timeExcludedMeta)))
+	m := newTestFetcherMetrics()
+	testutil.Ok(t, f.Filter(ctx, input, m.synced, false))
+
+	testutil.Equals(t, 2.0, promtest.ToFloat64(m.synced.WithLabelValues(timeExcludedMeta)))
 	testutil.Equals(t, expected, input)
 
 }
@@ -830,7 +847,7 @@ func TestDeduplicateFilter_Filter(t *testing.T) {
 	} {
 		f := NewDeduplicateFilter()
 		if ok := t.Run(tcase.name, func(t *testing.T) {
-			synced := promauto.With(nil).NewGaugeVec(prometheus.GaugeOpts{}, []string{"state"})
+			m := newTestFetcherMetrics()
 			metas := make(map[ulid.ULID]*metadata.Meta)
 			inputLen := len(tcase.input)
 			for id, metaInfo := range tcase.input {
@@ -848,12 +865,62 @@ func TestDeduplicateFilter_Filter(t *testing.T) {
 					},
 				}
 			}
-			testutil.Ok(t, f.Filter(ctx, metas, synced, false))
+			testutil.Ok(t, f.Filter(ctx, metas, m.synced, false))
 			compareSliceWithMapKeys(t, metas, tcase.expected)
-			testutil.Equals(t, float64(inputLen-len(tcase.expected)), promtest.ToFloat64(synced.WithLabelValues(duplicateMeta)))
+			testutil.Equals(t, float64(inputLen-len(tcase.expected)), promtest.ToFloat64(m.synced.WithLabelValues(duplicateMeta)))
 		}); !ok {
 			return
 		}
+	}
+}
+
+func TestReplicaLabelRemover_Modify(t *testing.T) {
+	ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
+	defer cancel()
+	rm := NewReplicaLabelRemover(log.NewNopLogger(), []string{"replica", "rule_replica"})
+
+	for _, tcase := range []struct {
+		name     string
+		input    map[ulid.ULID]*metadata.Meta
+		expected map[ulid.ULID]*metadata.Meta
+		modified float64
+	}{
+		{
+			name: "without replica labels",
+			input: map[ulid.ULID]*metadata.Meta{
+				ULID(1): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something"}}},
+				ULID(2): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something"}}},
+				ULID(3): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something1"}}},
+			},
+			expected: map[ulid.ULID]*metadata.Meta{
+				ULID(1): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something"}}},
+				ULID(2): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something"}}},
+				ULID(3): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something1"}}},
+			},
+			modified: 0,
+		},
+		{
+			name: "with replica labels",
+			input: map[ulid.ULID]*metadata.Meta{
+				ULID(1): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something"}}},
+				ULID(2): {Thanos: metadata.Thanos{Labels: map[string]string{"replica": "cluster1", "message": "something"}}},
+				ULID(3): {Thanos: metadata.Thanos{Labels: map[string]string{"replica": "cluster1", "rule_replica": "rule1", "message": "something"}}},
+				ULID(4): {Thanos: metadata.Thanos{Labels: map[string]string{"replica": "cluster1", "rule_replica": "rule1"}}},
+			},
+			expected: map[ulid.ULID]*metadata.Meta{
+				ULID(1): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something"}}},
+				ULID(2): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something"}}},
+				ULID(3): {Thanos: metadata.Thanos{Labels: map[string]string{"message": "something"}}},
+				ULID(4): {Thanos: metadata.Thanos{Labels: map[string]string{}}},
+			},
+			modified: 5.0,
+		},
+	} {
+		m := newTestFetcherMetrics()
+		testutil.Ok(t, rm.Modify(ctx, tcase.input, m.modified, false))
+
+		testutil.Equals(t, tcase.modified, promtest.ToFloat64(m.modified.WithLabelValues(replicaRemovedMeta)))
+		testutil.Equals(t, tcase.expected, tcase.input)
 	}
 }
 
@@ -945,7 +1012,7 @@ func TestConsistencyDelayMetaFilter_Filter_0(t *testing.T) {
 	}
 
 	t.Run("consistency 0 (turned off)", func(t *testing.T) {
-		synced := promauto.With(nil).NewGaugeVec(prometheus.GaugeOpts{}, []string{"state"})
+		m := newTestFetcherMetrics()
 		expected := map[ulid.ULID]*metadata.Meta{}
 		// Copy all.
 		for _, id := range u.created {
@@ -956,13 +1023,13 @@ func TestConsistencyDelayMetaFilter_Filter_0(t *testing.T) {
 		f := NewConsistencyDelayMetaFilter(nil, 0*time.Second, reg)
 		testutil.Equals(t, map[string]float64{"consistency_delay_seconds": 0.0}, extprom.CurrentGaugeValuesFor(t, reg, "consistency_delay_seconds"))
 
-		testutil.Ok(t, f.Filter(ctx, input, synced, false))
-		testutil.Equals(t, 0.0, promtest.ToFloat64(synced.WithLabelValues(tooFreshMeta)))
+		testutil.Ok(t, f.Filter(ctx, input, m.synced, false))
+		testutil.Equals(t, 0.0, promtest.ToFloat64(m.synced.WithLabelValues(tooFreshMeta)))
 		testutil.Equals(t, expected, input)
 	})
 
 	t.Run("consistency 30m.", func(t *testing.T) {
-		synced := promauto.With(nil).NewGaugeVec(prometheus.GaugeOpts{}, []string{"state"})
+		m := newTestFetcherMetrics()
 		expected := map[ulid.ULID]*metadata.Meta{}
 		// Only certain sources and those with 30m or more age go through.
 		for i, id := range u.created {
@@ -981,8 +1048,8 @@ func TestConsistencyDelayMetaFilter_Filter_0(t *testing.T) {
 		f := NewConsistencyDelayMetaFilter(nil, 30*time.Minute, reg)
 		testutil.Equals(t, map[string]float64{"consistency_delay_seconds": (30 * time.Minute).Seconds()}, extprom.CurrentGaugeValuesFor(t, reg, "consistency_delay_seconds"))
 
-		testutil.Ok(t, f.Filter(ctx, input, synced, false))
-		testutil.Equals(t, float64(len(u.created)-len(expected)), promtest.ToFloat64(synced.WithLabelValues(tooFreshMeta)))
+		testutil.Ok(t, f.Filter(ctx, input, m.synced, false))
+		testutil.Equals(t, float64(len(u.created)-len(expected)), promtest.ToFloat64(m.synced.WithLabelValues(tooFreshMeta)))
 		testutil.Equals(t, expected, input)
 	})
 }
@@ -1033,9 +1100,9 @@ func TestIgnoreDeletionMarkFilter_Filter(t *testing.T) {
 			ULID(4): {},
 		}
 
-		synced := promauto.With(nil).NewGaugeVec(prometheus.GaugeOpts{}, []string{"state"})
-		testutil.Ok(t, f.Filter(ctx, input, synced, false))
-		testutil.Equals(t, 1.0, promtest.ToFloat64(synced.WithLabelValues(markedForDeletionMeta)))
+		m := newTestFetcherMetrics()
+		testutil.Ok(t, f.Filter(ctx, input, m.synced, false))
+		testutil.Equals(t, 1.0, promtest.ToFloat64(m.synced.WithLabelValues(markedForDeletionMeta)))
 		testutil.Equals(t, expected, input)
 	})
 }